requirements.txt
fastapi
uvicorn
python-multipart
httpx
openai
python-dotenv
pydantic
google-genai
